Support delete requests
Add support to make delete requests to the ADP API so that consumers
may make requests to delete resources. Also includes some backwards
compatiblity breaking changes so that we can make raw get requests
that will return the whole response and not just the parsed response.
This was needed so that we can access the response headers for the
message id when querying the event notification message end points.

Closes #2